Set state in RegisterForm
