+++ src/Components/App.js
@@ -10,37 +10,6 @@ import ZipCodeForm from '../Components/ZipCodeForm.js'
 import LandingPage from '../Components/LandingPage.js'
 import { useQuery, gql } from '@apollo/client'
 
-const CREATE_LOCATION = gql`
-	mutation CreateLocation($email: String!, $streetAddress: String!, $city: String!, $state: String!, $zipcode: Int!, $locationType: String!, $scarinessLevel: Int!, $description: String, $startTime: String!, $endTime: String!, $image: String) {
-		createLocation(input: {
-			email: $email,
-			streetAddress: $streetAddress,
-			city: $city,
-			state: $state,
-			zipcode: $zipcode,
-			locationType: $locationType,
-			scarinessLevel: $scarinessLevel,
-			description: $description,
-			startTime: $startTime,
-			endTime: $endTime,
-			image: $image
-		}) {
-			location {
-				streetAddress
-        city
-        state
-        zipcode
-        locationType
-        scarinessLevel
-        description
-        startTime
-        endTime
-        image
-			}
-		}
-	}	
-`;
-
 const App = () => {
